genAi hachathon/backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
google-search-results
